src/main/webapp/app/entities/hersteller/hersteller.service.ts
import axios from 'axios';

import { IHersteller } from '@/shared/model/hersteller.model';

const baseApiUrl = 'api/herstellers';
const baseSearchApiUrl = 'api/_search/herstellers?query=';

export default class HerstellerService {
  public search(query): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(`${baseSearchApiUrl}${query}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public find(id: number): Promise<IHersteller> {
    return new Promise<IHersteller>(resolve => {
      axios.get(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public retrieve(): Promise<any> {
    return new Promise<any>(resolve => {
      axios.get(baseApiUrl).then(function(res) {
        resolve(res);
      });
    });
  }

  public delete(id: number): Promise<any> {
    return new Promise<any>(resolve => {
      axios.delete(`${baseApiUrl}/${id}`).then(function(res) {
        resolve(res);
      });
    });
  }

  public create(entity: IHersteller): Promise<IHersteller> {
    return new Promise<IHersteller>(resolve => {
      axios.post(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }

  public update(entity: IHersteller): Promise<IHersteller> {
    return new Promise<IHersteller>(resolve => {
      axios.put(`${baseApiUrl}`, entity).then(function(res) {
        resolve(res.data);
      });
    });
  }
}
